requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.3
pandas==2.2.3
pip==24.0
praw==7.8.1
prawcore==2.4.0
pysolr==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
tzdata==2025.1
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
flask